GSVA analysis of proteomics data
This commit includes the code to run GSVA analysis on the whole cell proteomics data for the HOPE cohort. The GSVA R library is implemented to run the GSVA enrichment analysis. The Kids First IDs are changed to sample IDs (SG IDs) in order to compare to other data and established cell lines.